Include and maintain input event codes
The previous dynamic build-time mechanism for generating ev.py was
just a recipe for troubles. The recent change in the filepath of the
event code header file is just one example.

Dynamic update worked pretty well, but because Linux is known for its
uapi stability, we can safely rely on them to maintain event codes in
backwards-compatible manner. After all, Python-uinput is just a
wrapper for uinput.

src/ev.py should be updated whenever new event codes are introduced in
mainline kernels. Yes, it is a bit more work, but it's a lot simpler
code-wise and the update process is trivial to (semi-)automate.